+++ example/lib/toast_no_context.dart
@@ -14,6 +14,7 @@ class ToastNoContext extends StatelessWidget {
       msg: "This is Colored Toast with android duration of 5 Sec",
       toastLength: Toast.LENGTH_SHORT,
       webBgColor: "#e74c3c",
+      textColor: Colors.black,
       timeInSecForIosWeb: 5,
     );
   }